services/seo-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis==5.0.4
arq==0.26.1
alembic==1.13.1
openai==1.30.1
structlog>=24.1.0
httpx==0.27.0
